# specs/openapi.baseline.json
{
  "components": {
...
    "responses": {
      "Error400": {
        "content": {
          "application/json": {
            "examples": {
              "bad_request": {
                "value": {
                  "error": "bad_request",
                  "message": "Bad Request"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Bad Request"
      },
      "Error401": {
        "content": {
          "application/json": {
            "examples": {
              "unauthorized": {
                "value": {
                  "error": "unauthorized",
                  "message": "Bearer token missing or invalid"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Unauthorized"
      },
      "Error403": {
        "content": {
          "application/json": {
            "examples": {
              "forbidden": {
                "value": {
                  "error": "forbidden",
                  "message": "Requires role admin",
                  "required_role": "admin"
                }
              }
            },
            "schema": {
              "properties": {
                "error": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "required_role": {
                  "type": "string"
                }
              },
              "required": [
                "error",
                "message"
              ],
              "type": "object"
            }
          }
        },
        "description": "Forbidden"
      },
      "Error404": {
        "content": {
          "application/json": {
            "examples": {
              "not_found": {
                "value": {
                  "error": "not_found",
                  "message": "Not Found"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Not Found"
      },
      "Error409": {
        "content": {
          "application/json": {
            "examples": {
              "conflict": {
                "value": {
                  "error": "conflict",
                  "message": "Conflict"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Conflict"
      },
...
      "Error500": {
        "content": {
          "application/json": {
            "examples": {
              "internal_error": {
                "value": {
                  "error": "internal_error",
                  "message": "Internal Server Error"
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Internal Server Error"
      },
...
    },
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "example": "validation_error",
            "type": "string"
          },
          "message": {
            "example": "Invalid input",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "pages"
        ],
        "type": "object"
      },
...
      "PageResponse_Tasks": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "items",
          "meta"
        ],
        "type": "object"
      },
      "Task": {
        "properties": {
          "assignee_id": {
            "nullable": true,
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "creator_user_id": {
            "nullable": true,
            "type": "integer"
          },
          "dish_id": {
            "nullable": true,
            "type": "integer"
          },
          "done": {
            "description": "Derived legacy boolean (status==\"done\"). Prefer using status for writes.",
            "readOnly": true,
            "type": "boolean"
          },
          "id": {
            "type": "integer"
          },
          "menu_id": {
            "nullable": true,
            "type": "integer"
          },
          "private_flag": {
            "type": "boolean"
          },
          "status": {
            "description": "Authoritative state. On create/update maps to done=(status==done).",
            "enum": [
              "todo",
              "doing",
              "blocked",
              "done",
              "cancelled"
            ],
            "type": "string"
          },
          "task_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "unit_id": {
            "nullable": true,
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "task_type",
          "done"
        ],
        "type": "object"
      },
      "TaskCreate": {
        "properties": {
          "private_flag": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "todo",
          "doing",
          "blocked",
          "done",
          "cancelled"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Unified Platform API",
    "version": "0.3.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/tasks/": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_Tasks"
                }
              }
            },
            "description": "Paged tasks list"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created task"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Delete task"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Get task"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "put": {
        "responses": {
          "200": {
            "description": "Update task"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}